split utils in utils.cc & utils.h
