+++ c_src/erocksdb.cc
@@ -58,6 +58,10 @@
 
 #include "detail.hpp"
 
+#ifndef INCL_UTIL_H
+    #include "util.h"
+#endif
+
 static ErlNifFunc nif_funcs[] =
 {
     // db operations
@@ -257,30 +261,6 @@ ERL_NIF_TERM ATOM_WRITE_THREADS;
 using std::nothrow;
 
 struct erocksdb_itr_handle;
-
-
-
-// Erlang helpers:
-ERL_NIF_TERM error_einval(ErlNifEnv* env)
-{
-    return enif_make_tuple2(env, erocksdb::ATOM_ERROR, erocksdb::ATOM_EINVAL);
-}
-
-static ERL_NIF_TERM error_tuple(ErlNifEnv* env, ERL_NIF_TERM error, rocksdb::Status& status)
-{
-    ERL_NIF_TERM reason = enif_make_string(env, status.ToString().c_str(),
-                                           ERL_NIF_LATIN1);
-    return enif_make_tuple2(env, erocksdb::ATOM_ERROR,
-                            enif_make_tuple2(env, error, reason));
-}
-
-static ERL_NIF_TERM slice_to_binary(ErlNifEnv* env, rocksdb::Slice s)
-{
-    ERL_NIF_TERM result;
-    unsigned char* value = enif_make_new_binary(env, s.size(), &result);
-    memcpy(value, s.data(), s.size());
-    return result;
-}
 
 ERL_NIF_TERM parse_db_option(ErlNifEnv* env, ERL_NIF_TERM item, rocksdb::Options& opts)
 {
